+++ src/components/base/new_dropdowns/base_dropdown/base_dropdown.vue
@@ -9,7 +9,7 @@ import {
   dropdownVariantOptions,
 } from '../../../../utils/constants';
 import {
-  GL_DROPDOWN_BOUNDARY_SELECTOR,
+  GL_DROPDOWN_HORIZONTAL_BOUNDARY_SELECTOR,
   GL_DROPDOWN_SHOWN,
   GL_DROPDOWN_HIDDEN,
   GL_DROPDOWN_BEFORE_CLOSE,
@@ -21,7 +21,12 @@ import {
   POSITION_ABSOLUTE,
   POSITION_FIXED,
 } from '../constants';
-import { logWarning, isElementTabbable, isElementFocusable } from '../../../../utils/utils';
+import {
+  logWarning,
+  isElementTabbable,
+  isElementFocusable,
+  getHorizontalBoundingClientRect,
+} from '../../../../utils/utils';
 
 import GlButton from '../../button/button.vue';
 import GlIcon from '../../icon/icon.vue';
@@ -271,10 +276,15 @@ export default {
         strategy: this.positioningStrategy,
         middleware: [
           offset(this.offset),
-          autoPlacement({
-            alignment,
-            boundary: this.$el.closest(GL_DROPDOWN_BOUNDARY_SELECTOR) || 'clippingAncestors',
-            allowedPlacements: dropdownAllowedAutoPlacements[this.placement],
+          autoPlacement(() => {
+            const autoHorizontalBoundary = getHorizontalBoundingClientRect(
+              this.$el.closest(GL_DROPDOWN_HORIZONTAL_BOUNDARY_SELECTOR)
+            );
+            return {
+              alignment,
+              boundary: autoHorizontalBoundary || 'clippingAncestors',
+              allowedPlacements: dropdownAllowedAutoPlacements[this.placement],
+            };
           }),
           shift(),
           size({