+++ src/components/experimental/duo/chat/components/duo_chat_message/duo_chat_message.stories.js
@@ -33,6 +33,14 @@ Response.args = generateProps({
   message: MOCK_RESPONSE_MESSAGE,
 });
 
+export const ErrorResponse = Template.bind({});
+ErrorResponse.args = generateProps({
+  message: {
+    ...MOCK_RESPONSE_MESSAGE,
+    errors: ['Error: Whatever you see is wrong'],
+  },
+});
+
 export default {
   title: 'experimental/duo/chat/duo-chat-message',
   component: GlDuoChatMessage,