+++ src/tokens/text.default.tokens.stories.js
@@ -2,7 +2,12 @@ import COMPILED_TOKENS from '../../dist/tokens/json/tokens.json';
 import { createDesignTokenStory } from './common_story_options';
 
 export const Default = createDesignTokenStory({
-  tokens: COMPILED_TOKENS.text,
+  tokens: {
+    primary: COMPILED_TOKENS.text.primary,
+    secondary: COMPILED_TOKENS.text.secondary,
+    tertiary: COMPILED_TOKENS.text.tertiary,
+    ...COMPILED_TOKENS.text.color,
+  },
   isBackgroundColorStory: false,
 });